issue 1152 - import access roles as objects with own object type
