+++ roles/importer/files/importer/checkpointR8x/cp_network.py
@@ -84,8 +84,8 @@ def collect_nw_objects(object_table, nw_objects, debug_level=0, mgm_id=0):
                         else:
                             obj['uid'] = obj['name-in-updatable-objects-repository']
                         obj['color'] = 'black'
-                    # TODO: handle exclusion groups, access-roles correctly
-                    if obj_type in ['updatable-object', 'access-role', 'group-with-exclusion', 'security-zone', 'dns-domain']:
+                        
+                    if obj_type in ['updatable-object', 'group-with-exclusion', 'security-zone', 'dns-domain']:
                         obj_type = 'group'
 
                     if obj_type == 'group-with-exclusion':
