+++ roles/lib/files/FWO.Basics/GlobalConstants.cs
@@ -62,6 +62,7 @@ namespace FWO.Basics
         public const string Host = "host";
         public const string Network = "network";
         public const string IPRange = "ip_range";
+        public const string AccessRole = "access-role";
     }
 
     public struct ServiceType
